+++ include/core/parameters.h
@@ -296,6 +296,8 @@ namespace Parameters
     // Properties that apply to the tracer physics with immersed solids
     double tracer_diffusivity_inside;
     double tracer_diffusivity_outside;
+    double tracer_reaction_constant_inside;
+    double tracer_reaction_constant_outside;
     double thickness;
 
     static void
@@ -430,6 +432,11 @@ namespace Parameters
     // tracer diffusivity in L^2/s
     double tracer_diffusivity;
 
+    // tracer reaction constant in 1/s^[order]
+    double tracer_reaction_constant;
+    // tracer reaction order
+    double tracer_reaction_order;
+
     // Phase change parameters
     PhaseChange phase_change_parameters;
 
@@ -474,6 +481,13 @@ namespace Parameters
       constant,
       immersed_boundary_tanh
     } tracer_diffusivity_model;
+
+    enum class TracerReactionPrefactorModel
+    {
+      none,
+      constant,
+      immersed_boundary_tanh
+    } tracer_reaction_prefactor_model;
 
     // Struct that contains the parameters to handle physical properties when
     // immersed solids models are used